Add support for config files
Two main things:

(1) Teach CLI to accept `--config` option.

If provided, we try to parse the file as JSON, and we insert the
key/value pairs into our options object. Arguments passed in on the
command-line take precedence over values coming in from the config file.

(2) Look for config file in standard locations if none is provided.

We first look for ".jscodeshiftrc" in the current directory, falling
back to ".jscodeshiftrc" in the `HOME` and then finally to
"/etc/jscodeshiftrc" (or the equivalent on Windows).

Notes:

- Created a `logError` module to keep things DRY along the way.
- Created `OptionParser`, `ConfigReader` and `Process` modules for
  testability, and to make it possible to isolate tests from
  system-local configuration by mocking.

Fixes:

  https://github.com/facebook/jscodeshift/issues/4